Extract scoring/timer math into src/scoring.rs
Lifts six pure functions and their tests out of main.rs into a new
library module. No behavior change.

Moves: TIME_BONUS_SCORE_PER_SECOND, level_time_limit_duration,
ceil_duration_seconds, format_level_timer, time_bonus_score,
level_index_for_progress, combined_score — and their four tests.

main.rs drops from 1576 → 1502 lines.
